Implement process_sync_committee

+++ consensus/state_processing/src/common/initiate_validator_exit.rs
@@ -9,7 +9,7 @@ pub fn initiate_validator_exit<T: EthSpec>(
     spec: &ChainSpec,
 ) -> Result<(), Error> {
     if index >= state.validators().len() {
-        return Err(Error::UnknownValidator(index as u64));
+        return Err(Error::UnknownValidator(index));
     }
 
     // Return if the validator already initiated exit
